Listar Clientes e Adicionar Clientes com thymeleaf

+++ apiksi/src/main/java/com/claudiathalita/apiksi/repository/ClienteRepository.java
@@ -3,9 +3,10 @@ package com.claudiathalita.apiksi.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.claudiathalita.apiksi.Cliente;
+import com.claudiathalita.apiksi.entity.Cliente;
+
 
 @Repository
-public interface ClienteRepository extends JpaRepository<Cliente, Integer>{
+public interface ClienteRepository extends JpaRepository<Cliente, Long>{
 
 }
